Splitter opp PdlClient.kt og PdlClientImpl.kt
